+++ clipspyx/dsl/__init__.py
@@ -1,6 +1,7 @@
 from clipspyx.dsl.types import Multi, Fact
 from clipspyx.dsl.template import Template
 from clipspyx.dsl.rule import Rule, _Placeholder
+from clipspyx.dsl.timer import TimerEvent, AFTER, AT, EVERY
 
 
 def before(target):
@@ -20,4 +21,5 @@ def concurrent(target):
 
 NIL = None
 
-__all__ = ['Template', 'Rule', 'Multi', 'Fact', 'NIL', 'before', 'after', 'concurrent']
+__all__ = ['Template', 'Rule', 'Multi', 'Fact', 'NIL', 'before', 'after', 'concurrent',
+           'TimerEvent', 'AFTER', 'AT', 'EVERY']